backend/requirements.txt
asgiref==3.11.1
djangorestframework>=3.15
django-cors-headers>=4.0
djangorestframework-simplejwt>=5.0
diff-match-patch==20241021
Django==6.0.2
django-admin-filters==0.1.3
django-admin-list-filter-dropdown==1.0.3
django-autocomplete-light==3.12.1
django-filter==25.2
django-import-export==4.4.0
django-jazzmin==3.0.2
django-js-asset==3.1.2
django-lineup==1.0.0
django-mptt==0.18.0
django-smart-selects==1.7.2
et_xmlfile==2.0.0
numpy==2.4.3
openpyxl==3.1.5
pandas==3.0.1
pillow==12.1.1
psycopg==3.3.2
psycopg-binary==3.3.2
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.5
tablib==3.9.0
tzdata==2025.3
